+++ app/src/main/res/layout/activity_add_income_expense.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AddIncomeExpenseActivity">
+    tools:context=".ui.add_money_transaction.AddIncomeExpenseActivity">
 
     <TextView
         android:id="@+id/income_expense_title"
